Feature: Add check script
