+++ package.json
@@ -9,7 +9,8 @@
     "test": "jest",
     "lint": "eslint .",
     "lint:fix": "eslint --fix . --config .eslintrc.js --ext .js",
-    "format": "prettier --write ."
+    "format": "prettier --write .",
+    "check": "npm run test && npm run lint"
   },
   "author": "Justin Looi",
   "license": "ISC",
